StatelessFeature: Rename to AlwaysOnFeature
