+++ common/src/main/java/com/wynntils/core/features/FeatureRegistry.java
@@ -9,10 +9,10 @@ import com.wynntils.core.config.ConfigManager;
 import com.wynntils.core.features.properties.RegisterKeyBind;
 import com.wynntils.core.features.properties.StartDisabled;
 import com.wynntils.core.keybinds.KeyHolder;
+import com.wynntils.features.alwayson.FixSpellOverwriteFeature;
+import com.wynntils.features.alwayson.LootrunFeature;
 import com.wynntils.features.debug.ConnectionProgressFeature;
 import com.wynntils.features.debug.PacketDebuggerFeature;
-import com.wynntils.features.stateless.FixSpellOverwriteFeature;
-import com.wynntils.features.stateless.LootrunFeature;
 import com.wynntils.features.user.ChatItemFeature;
 import com.wynntils.features.user.DialogueOptionOverrideFeature;
 import com.wynntils.features.user.EmeraldPouchHotkeyFeature;
