+++ common/src/main/java/com/wynntils/features/alwayson/LootrunFeature.java
@@ -2,12 +2,12 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.features.stateless;
+package com.wynntils.features.alwayson;
 
 import com.google.common.collect.ImmutableList;
 import com.wynntils.core.config.Config;
 import com.wynntils.core.config.ConfigHolder;
-import com.wynntils.core.features.StatelessFeature;
+import com.wynntils.core.features.AlwaysOnFeature;
 import com.wynntils.core.features.properties.StartDisabled;
 import com.wynntils.core.managers.Manager;
 import com.wynntils.mc.event.ClientTickEvent;
@@ -24,7 +24,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 @StartDisabled
-public class LootrunFeature extends StatelessFeature {
+public class LootrunFeature extends AlwaysOnFeature {
     public static LootrunFeature INSTANCE;
 
     // TODO: Add textured path type
